api/pyproject.toml
[project]
name = "data-agent-api"
version = "0.1.0"
description = "FastAPI server with Microsoft Agent Framework with support for Microsoft Foundry agents. Uses SSE for streaming responses."
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "agent-framework>=1.0.0b251223",
    "agent-framework-devui>=1.0.0b251223",
    "aioodbc>=0.5.0",
    "azure-identity>=1.15.0",
    "azure-search-documents>=11.5.0",
    "fastapi>=0.115.0",
    "fastapi-azure-auth>=5.0.0",
    "openai>=1.0.0",
    "uvicorn[standard]>=0.30.0",
    "python-dotenv>=1.0.0",
    "pydantic>=2.0.0",
]

[project.optional-dependencies]
observability = [
    "azure-monitor-opentelemetry>=1.0.0",
    "opentelemetry-exporter-otlp-proto-grpc>=1.20.0",
]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "httpx>=0.27.0",
]

[tool.pyright]
pythonVersion = "3.12.3"
include = ["src"]
extraPaths = ["."]

[tool.pylint.MASTER]
# No special path hooks needed - src is a proper package